move store.js to store/s3.js
in order to allow different storage backends
