+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "websocket.js",
   "scripts": {
-    "test": "eslint *.js database/*.js && node test/client.js",
+    "test": "eslint *.js database/*.js store/*.js && node test/client.js",
     "start": "NODE_ENV=production node app.js",
     "postinstall": "wget http://geolite.maxmind.com/download/geoip/database/GeoLite2-City.mmdb.gz && gunzip GeoLite2-City.mmdb.gz"
   },
